+++ Assets/Xsolla/Core/Browser/XsollaWebBrowser.cs
@@ -89,7 +89,7 @@ namespace Xsolla.Core
 				};
 
 				Screen.fullScreen = false;
-				OpenPaystationWidget(paymentToken, XsollaSettings.IsSandbox);
+				OpenPaystationWidget(paymentToken, XsollaSettings.IsSandbox, Application.unityVersion, Constants.SDK_VERSION);
 				return;
 			}
 #endif
@@ -100,7 +100,6 @@ namespace Xsolla.Core
 #if UNITY_STANDALONE || UNITY_EDITOR
 			if (InAppBrowser != null && !forcePlatformBrowser)
 			{
-				InAppBrowser.AddInitHandler(() => InAppBrowser.UpdateSize(740, 760));
 				InAppBrowser.CloseEvent += onBrowserClosed;
 			}
 #endif
@@ -136,7 +135,7 @@ namespace Xsolla.Core
 
 #if UNITY_WEBGL
 		[DllImport("__Internal")]
-		private static extern void OpenPaystationWidget(string token, bool sandbox);
+		private static extern void OpenPaystationWidget(string token, bool sandbox, string engineVersion, string sdkVersion);
 
 		[DllImport("__Internal")]
 		private static extern void ClosePaystationWidget();